move handlers from index.ts in seperate file
